转载+转载统计

+++ models/post.js
@@ -21,6 +21,7 @@ Post.prototype.save = function (callback) {
         head: this.head,
         post: this.post,
         tag: this.tag,
+        reprint_info: JSON.stringify({}),
         pv: 0,
         comments: JSON.stringify([])
     }
@@ -116,7 +117,7 @@ Post.getOne = function (query, callback) {
             console.log('error: ', error)
             return callback(error);
         }
-        if (results) {
+        if (results && results.length > 0) {
             results.forEach(doc => {
                 doc.post = markdown.toHTML(doc.post);
                 doc.comments = JSON.parse(doc.comments);
@@ -134,6 +135,8 @@ Post.getOne = function (query, callback) {
                 console.log(update.sql);
             }
             return callback(null, results);
+        } else {
+            return callback('它消失好久了!')
         }
     });
     console.log(findOne.sql)
@@ -167,12 +170,6 @@ Post.edit = function (query, callback) {
 
 Post.update = function (query, callback) {
     if (query.name && query.time && query.title) {
-        /**
-         *connection.query('UPDATE users SET foo = ?, bar = ?, baz = ? WHERE id = ?', ['a', 'b', 'c', userId], function (error, results, fields) {
-         *    if (error) throw error;
-         *    // ...
-         *});
-         */
         let sqlString = 'update t_blog_post set ? where user_name = ? and create_time = ? ';
         let params = {
             user_name: query.name,
@@ -193,20 +190,74 @@ Post.update = function (query, callback) {
 
 Post.remove = function (query, callback) {
     if (query.name && query.time && query.title) {
-        let sqlString = 'delete from t_blog_post where user_name = ? and create_time = ? and title = ? ';
-        let remove = conn.query(sqlString, [query.name, query.time, query.title], function (error, results, fields) {
+        //1、获取该文章信息,判断是否存在reprint_from
+        query.aginAsk = true;
+        let sqlString = '';
+        Post.getOne(query, function(error, posts) {
             if (error) {
                 console.log('error: ', error)
                 return callback(error);
             }
-            if (results.affectedRows > 0) return callback(null, 'success');
-            return callback(null);
-        });
-        console.log(remove.sql);
+            if(posts.length > 0) {
+                let origin_from_info = JSON.parse(posts[0].reprint_info);
+                //2、存在reprint_from，说明该文章为转载， 更新源文件reprint_to记录
+                if(origin_from_info.reprint_from) {
+                    let origin_from = origin_from_info.reprint_from;
+                    sqlString = 'select * from t_blog_post where user_name = ? and create_time = ? and title = ? ';
+                    let find = conn.query(sqlString, [origin_from.user_name, origin_from.create_time, origin_from.title], function (error, results, fields) {
+                        if (error) {
+                            console.log('error: ', error)
+                            return callback(error);
+                        }
 
+                        if(results.length > 0) {
+                            let post = results[0];
+                            //3、获取原文章的reprint_to，删除对应的记录
+                            let origin_to = JSON.parse(post.reprint_info).reprint_to;
+                            for(var i = 0; i < origin_to.length; i++) {
+                                let item = origin_to[i];
+                                if(item.user_name == query.name && item.create_time == query.time && item.title == query.title) {
+                                    console.log('i: ', i)
+                                    origin_to.splice(i, 1);
+                                }
+                            }
+                            sqlString = 'update t_blog_post set ? where user_name = ? and create_time = ? ';
+                            let params = {reprint_info: JSON.stringify({"reprint_to": origin_to})};
+                            let update = conn.query(sqlString, [params, origin_from.user_name, origin_from.create_time], function (error, results, fields) {
+                                if (error) {
+                                    console.log('error: ', error)
+                                    return callback(error);
+                                }
+                            });
+                            console.log(update.sql);
+                        }
+
+                    });
+                    console.log(find.sql);
+                }
+
+                //4、最后删除文章操作
+                sqlString = 'delete from t_blog_post where user_name = ? and create_time = ? and title = ? ';
+                let remove = conn.query(sqlString, [query.name, query.time, query.title], function (error, results, fields) {
+                    if (error) {
+                        console.log('error: ', error)
+                        return callback(error);
+                    }
+                    if (results.affectedRows > 0) return callback(null, 'success');
+                    return callback(null);
+                });
+                console.log(remove.sql);
+
+            }
+
+        });
+
+    } else {
+        return callback('参数异常！');
     }
 
 }
+
 /**
  * 根据用户名获取标签列表
  * @param {*} name 
@@ -278,7 +329,65 @@ Post.search = function (keyWord, callback) {
             return callback(null, results);
         }
     });
-
     console.log(findFuzzys.sql)
+}
+
+Post.reprint = function(reprint_from, reprint_to, callback) {
+    //1、找到原始文章
+    let query = reprint_from;
+    query.aginAsk = true;
+    Post.getOne(query, (err, posts) => {
+        if (err) {
+            return callback(err);
+        }
+        if(posts && posts.length > 0) {
+            let post = posts[0];
+            let doc = {};
+            doc.user_name = reprint_to.name;
+            doc.head = reprint_to.head;
+            doc.create_time = new Date().getTime();
+            doc.title = (post.title.search(/[转载]/) > -1) ? post.title : "[转载]" + post.title;
+            doc.post = post.post;
+            let from = {user_name: reprint_from.name, create_time: reprint_from.time, title:reprint_from.title};
+            doc.reprint_info = JSON.stringify({"reprint_from": from});
+            doc.comments = JSON.stringify([]);
+            doc.pv = 0;
+
+            // 2、更新原始文章的reprint_to
+            let sqlString = 'update t_blog_post set ? where user_name = ? and create_time = ? ';
+            let to = {user_name: reprint_to.name, create_time: doc.create_time, title: doc.title};
+            let info = JSON.parse(post.reprint_info);
+            let to_tmp = undefined;
+            if(info.reprint_to) {
+                to_tmp = info.reprint_to;
+            } else {
+                to_tmp = []; 
+            }
+            to_tmp.push(to);
+            let params = {reprint_info: JSON.stringify({"reprint_to": to_tmp})};
+            let update = conn.query(sqlString, [params, reprint_from.name, reprint_from.time], function (error, results, fields) {
+                if (error) {
+                    console.log('error: ', error)
+                    return callback(error);
+                }
+                // 3、新增转载记录
+                let sqlString = 'insert into t_blog_post set ? '
+                let insert = conn.query(sqlString, doc, function (error, results, fields) {
+                    if (error) {
+                        console.log('error: ', error)
+                        return callback(error);
+                    }
+                    if (results.insertId) {
+                        return callback(null, doc);
+                    }
+                });
+                console.log(insert.sql)
+
+            });
+            console.log(update.sql)
+
+        }
+
+    });
 
 }
